refactor(sdcard): extract shared ScaleRawAnalogValues helper (#462)
The three SD-card parsers (protobuf, CSV, JSON) each carried a
byte-identical ScaleRawAnalogValues body plus the same doc remarks.
Extracted into internal SdCardAnalogScaling.ScaleRawAnalogValues,
which all three now delegate to. Pure refactor; no behavior change.

Closes #462
