+++ src/Daqifi.Core/Device/SdCard/SdCardCsvFileParser.cs
@@ -339,7 +339,7 @@ public sealed class SdCardCsvFileParser
             var (rowTimestamp, rawAnalogValues, digitalData, perChannelTimestamps) = parsed.Value;
 
             // Scale raw ADC values using device calibration
-            var analogValues = ScaleRawAnalogValues(rawAnalogValues, config);
+            var analogValues = SdCardAnalogScaling.ScaleRawAnalogValues(rawAnalogValues, config);
 
             // Reconstruct absolute timestamp using first channel timestamp
             var absoluteTime = baseTime;
@@ -453,41 +453,6 @@ public sealed class SdCardCsvFileParser
         {
             return null;
         }
-    }
-
-    /// <summary>
-    /// Scales raw ADC values to real voltage using device calibration data.
-    /// Formula: <c>raw / resolution * portRange * calM * internalScaleM + calB</c>
-    /// (see <see cref="Channel.AnalogScaling"/> — <c>calB</c> is an offset in volts and is
-    /// deliberately not scaled by <c>internalScaleM</c>).
-    /// </summary>
-    private static IReadOnlyList<double> ScaleRawAnalogValues(
-        IReadOnlyList<double> rawValues,
-        SdCardDeviceConfiguration? config)
-    {
-        if (config == null || config.Resolution == 0)
-        {
-            // No config or resolution available — return raw values as-is
-            return rawValues;
-        }
-
-        var result = new double[rawValues.Count];
-        var resolution = (double)config.Resolution;
-        var cal = config.CalibrationValues;
-        var portRange = config.PortRange;
-        var intScale = config.InternalScaleM;
-
-        for (var ch = 0; ch < rawValues.Count; ch++)
-        {
-            var calM = cal != null && ch < cal.Count ? cal[ch].Slope : 1.0;
-            var calB = cal != null && ch < cal.Count ? cal[ch].Intercept : 0.0;
-            var range = portRange != null && ch < portRange.Count ? portRange[ch] : 1.0;
-            var scaleM = intScale != null && ch < intScale.Count ? intScale[ch] : 1.0;
-
-            result[ch] = Channel.AnalogScaling.Scale(rawValues[ch], resolution, range, calM, scaleM, calB);
-        }
-
-        return result;
     }
 
     /// <summary>
