+++ src/Daqifi.Core/Device/SdCard/SdCardFileParser.cs
@@ -304,14 +304,15 @@ public sealed class SdCardFileParser
             }
 
             // Extract analog values (prefer float, fall back to scaled raw int)
-            double[] analogValues;
+            IReadOnlyList<double> analogValues;
             if (msg.AnalogInDataFloat.Count > 0)
             {
                 analogValues = msg.AnalogInDataFloat.Select(v => (double)v).ToArray();
             }
             else if (msg.AnalogInData.Count > 0)
             {
-                analogValues = ScaleRawAnalogValues(msg.AnalogInData, config);
+                analogValues = SdCardAnalogScaling.ScaleRawAnalogValues(
+                    msg.AnalogInData.Select(v => (double)v).ToArray(), config);
             }
             else
             {
@@ -341,41 +342,6 @@ public sealed class SdCardFileParser
         }
     }
 #pragma warning restore CS1998
-
-    /// <summary>
-    /// Scales raw ADC integer values using calibration parameters from the device config.
-    /// Formula: <c>raw / resolution * portRange * calM * internalScaleM + calB</c>
-    /// (see <see cref="Channel.AnalogScaling"/> — <c>calB</c> is an offset in volts and is
-    /// deliberately not scaled by <c>internalScaleM</c>).
-    /// </summary>
-    private static double[] ScaleRawAnalogValues(
-        Google.Protobuf.Collections.RepeatedField<int> rawValues,
-        SdCardDeviceConfiguration? config)
-    {
-        if (config == null || config.Resolution == 0)
-        {
-            // No config or resolution available — return raw values as-is
-            return rawValues.Select(v => (double)v).ToArray();
-        }
-
-        var result = new double[rawValues.Count];
-        var resolution = (double)config.Resolution;
-        var cal = config.CalibrationValues; // May be null — defaults applied per-channel below
-        var portRange = config.PortRange;
-        var intScale = config.InternalScaleM;
-
-        for (var ch = 0; ch < rawValues.Count; ch++)
-        {
-            var calM = cal != null && ch < cal.Count ? cal[ch].Slope : 1.0;
-            var calB = cal != null && ch < cal.Count ? cal[ch].Intercept : 0.0;
-            var range = portRange != null && ch < portRange.Count ? portRange[ch] : 1.0;
-            var scaleM = intScale != null && ch < intScale.Count ? intScale[ch] : 1.0;
-
-            result[ch] = Channel.AnalogScaling.Scale(rawValues[ch], resolution, range, calM, scaleM, calB);
-        }
-
-        return result;
-    }
 
     /// <summary>
     /// Computes the tick delta between two uint32 timestamps, handling rollover.
